added styleable toasts.

+++ app/src/main/java/com/example/myapplication/changeFlightActivity.java
@@ -5,7 +5,9 @@ import androidx.room.Room;
 
 import android.content.Context;
 import android.content.Intent;
+import android.graphics.Color;
 import android.os.Bundle;
+import android.view.Gravity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -14,6 +16,8 @@ import android.widget.Toast;
 import com.example.myapplication.DB.AppDataBase;
 import com.example.myapplication.DB.FlightsDAO;
 import com.example.myapplication.databinding.ActivityChangeFlightBinding;
+
+import io.github.muddz.styleabletoast.StyleableToast;
 
 public class changeFlightActivity extends AppCompatActivity {
     private static final String CHANGE_FLIGHT_ACTIVITY_USER = "com.example.myapplication.changeFlightActivityUser";
@@ -71,7 +75,13 @@ public class changeFlightActivity extends AppCompatActivity {
 
                     Flights flight = new Flights(origin, destination, capacity);
                     mFlightsDAO.insert(flight);
-                    Toast.makeText(changeFlightActivity.this, "FLIGHT ADDED YAY!", Toast.LENGTH_SHORT).show();
+                    new StyleableToast
+                            .Builder(changeFlightActivity.this)
+                            .text("FLIGHT ADDED YAY!")
+                            .textColor(Color.WHITE)
+                            .gravity(Gravity.TOP)
+                            .backgroundColor(Color.GREEN)
+                            .show();
                 }
             }
         });
@@ -82,8 +92,24 @@ public class changeFlightActivity extends AppCompatActivity {
                 if (hasDeleteInputs()) {
                     int flightId = Integer.parseInt(mFlightID.getText().toString());
                     Flights flight = mFlightsDAO.getById(flightId);
-                    mFlightsDAO.delete(flight);
-                    Toast.makeText(changeFlightActivity.this, "FLIGHT DELETED!", Toast.LENGTH_SHORT).show();
+                    if (flight != null) {
+                        mFlightsDAO.delete(flight);
+                        new StyleableToast
+                                .Builder(changeFlightActivity.this)
+                                .text("FLIGHT DELETED!")
+                                .textColor(Color.WHITE)
+                                .gravity(Gravity.TOP)
+                                .backgroundColor(Color.GREEN)
+                                .show();
+                        return;
+                    }
+                    new StyleableToast
+                            .Builder(changeFlightActivity.this)
+                            .text("FLIGHT NOT FOUND!")
+                            .textColor(Color.WHITE)
+                            .gravity(Gravity.TOP)
+                            .backgroundColor(Color.RED)
+                            .show();
                 }
             }
         });
@@ -91,15 +117,33 @@ public class changeFlightActivity extends AppCompatActivity {
 
     public boolean hasAddInputs() {
         if (mOrigin.getText().toString().isEmpty()) {
-            Toast.makeText(changeFlightActivity.this, "Where From is Empty!", Toast.LENGTH_SHORT).show();
+            new StyleableToast
+                    .Builder(this)
+                    .text("Where From is Empty!")
+                    .textColor(Color.WHITE)
+                    .gravity(Gravity.TOP)
+                    .backgroundColor(Color.RED)
+                    .show();
             return false;
         }
         if (mDestination.getText().toString().isEmpty()) {
-            Toast.makeText(changeFlightActivity.this, "Where To is Empty!", Toast.LENGTH_SHORT).show();
+            new StyleableToast
+                    .Builder(this)
+                    .text("Where To is Empty!")
+                    .textColor(Color.WHITE)
+                    .gravity(Gravity.TOP)
+                    .backgroundColor(Color.RED)
+                    .show();
             return false;
         }
         if (mCapacity.getText().toString().isEmpty()) {
-            Toast.makeText(changeFlightActivity.this, "Capacity is Empty!", Toast.LENGTH_SHORT).show();
+            new StyleableToast
+                    .Builder(this)
+                    .text("Capacity is Empty!")
+                    .textColor(Color.WHITE)
+                    .gravity(Gravity.TOP)
+                    .backgroundColor(Color.RED)
+                    .show();
             return false;
         }
         return true;
@@ -107,7 +151,13 @@ public class changeFlightActivity extends AppCompatActivity {
 
     public boolean hasDeleteInputs() {
         if (mFlightID.getText().toString().isEmpty()) {
-            Toast.makeText(changeFlightActivity.this, "FlightID is Empty!", Toast.LENGTH_SHORT).show();
+            new StyleableToast
+                    .Builder(this)
+                    .text("FlightID is Empty!")
+                    .textColor(Color.WHITE)
+                    .gravity(Gravity.TOP)
+                    .backgroundColor(Color.RED)
+                    .show();
             return false;
         }
         return true;
